ai-sql-analyst/requirements.txt
streamlit
duckdb
python-dotenv
pandas
sqlparse
google-generativeai
pytest